chore: Build fixes
