+++ apps/garden/app/prijava/promjena-zaporke/page.tsx
@@ -1,0 +1,3 @@
+export default function ChangePasswordPage() {
+    return null;
+}
